[mod/#24] 컴포넌트 분리
