+++ feature/mypage/src/main/java/com/napzak/market/mypage/component/MyMarketButton.kt
@@ -1,0 +1,51 @@
+package com.napzak.market.mypage.component
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.Icon
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.vector.ImageVector
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.res.vectorResource
+import androidx.compose.ui.unit.dp
+import com.napzak.market.designsystem.R.drawable.ic_arrow_right_7
+import com.napzak.market.designsystem.theme.NapzakMarketTheme
+import com.napzak.market.feature.mypage.R.string.mypage_my_market
+import com.napzak.market.util.android.noRippleClickable
+
+@Composable
+fun MyMarketButton(onClick: () -> Unit) {
+    Row(
+        modifier = Modifier
+            .fillMaxWidth()
+            .clip(RoundedCornerShape(14.dp))
+            .background(NapzakMarketTheme.colors.gray10)
+            .padding(vertical = 14.dp)
+            .noRippleClickable(onClick),
+        verticalAlignment = Alignment.CenterVertically,
+        horizontalArrangement = Arrangement.Center,
+    ) {
+        Text(
+            text = stringResource(mypage_my_market),
+            style = NapzakMarketTheme.typography.caption12sb,
+            color = NapzakMarketTheme.colors.gray300,
+        )
+        Spacer(modifier = Modifier.width(6.dp))
+        Icon(
+            imageVector = ImageVector.vectorResource(ic_arrow_right_7),
+            contentDescription = null,
+            tint = Color.Unspecified,
+        )
+    }
+}
